Working on Gameboard and Pieces
Note that neither the gameboard nor pieces contain any game-rule logic. Gameboard is a basic container for pieces. I'm planning on having rules expert classes to handle all of of the rules. The gameboard allows 0 or 1 piece in each coordinate. Pieces have a color, and the type is determined by the class.