Friend Matchmaking
progress on #10

Choosing online multi-player will now show the options of random and friend again. Starting a matchmaking activity with the extra string named `opponentUsername` will initiate a matchmaking process targeting only a specified user. The same request as random matchmaking is sent to server but with an additional string added to the response named `username_of_opponent`.